requirements.txt
streamlit
pandas
matplotlib
python-pptx
openpyxl